+++ lib/widgets/widgets.dart
@@ -1,3 +1,3 @@
-export 'filter-widget.dart';
+export 'filter_widget.dart';
 export 'filter_checkbox_title.dart';
 export 'filter_text.dart';