+++ CHANGELOG.md
@@ -27,3 +27,22 @@
     15. dividerColor,
     16. dividerThickness,
     17. checkBoxTileThemeProps,
+
+
+## 0.0.4
+
+* Search property added.
+  ** Now user can make the local search for filter option with search decoration and other properties access.
+
+* Style properties for search view
+    1. OutlineInputBorder? inputBorder;
+    2. Color? fillColor;
+    3. Color? clearIconColor;
+    4. Color? searchIconColor;
+    5. Widget? clearIcon;
+    6. Widget? searchIcon;
+    7. bool? filled;
+    8. String? searchHint;
+    9. TextStyle? textStyle;
+    10. TextStyle? hintStyle;
+